Разделил тесты на модули
