+++ superlists/lists/tests/test_views.py
@@ -11,23 +11,6 @@ class HomePageTest(TestCase):
         response = self.client.get('/')
         self.assertTemplateUsed(response, 'home.html')
 
-    def test_only_save_items_when_necessary(self):
-        self.client.get('/')
-        self.assertEqual(Item.objects.count(), 0)
-
-    def test_can_save_a_POST_request(self):
-        '''тест: можно сохранить post-запрос'''
-        self.client.post('/', data={'item_text': 'A new list item'})
-        self.assertEqual(Item.objects.count(), 1)
-        new_item = Item.objects.first()
-        self.assertEqual(new_item.text, 'A new list item')
-
-    def test_redirects_after_POST(self):
-        '''тест: переадресация после post-запроса'''
-        response = self.client.post('/', data={'item_text': 'A new list item'})
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response['location'], '/')
-
     def test_display_all_list_items(self):
         '''тест: проверка отображения всех элементов списка'''
         Item.objects.create(text='itemey 1')
